remove duplicate and rename parent class
