+++ libs/ngx-slice-kit/src/lib/navigation/nav-tabs/nav-tabs.component.ts
@@ -1,23 +1,19 @@
 import {
     AfterContentInit,
-    AfterViewChecked,
-    AfterViewInit,
     ChangeDetectorRef,
     Component,
     ContentChildren,
-    Input,
-    OnDestroy,
     OnInit,
     QueryList,
     ViewChild
 } from '@angular/core';
 import { NavigationEnd, Router, RouterOutlet } from '@angular/router';
-import { BehaviorSubject, fromEvent, Subscription } from 'rxjs';
+import { fromEvent } from 'rxjs';
 import { delay, filter, throttleTime } from 'rxjs/operators';
 import { ThemeService } from '../../core/theme/theme.service';
 import { slideInAnimation } from '../../core/animations/slide-in';
 import { TabLinkDirective } from './tab-link.directive';
-import { BasicTabComponent } from '../basic-tab/basic-tab.component';
+import { TabsGroupComponent } from '../tabs-group/tabs-group.component';
 
 @Component({
     selector: 'sdk-nav-tab-group',
@@ -27,39 +23,16 @@ import { BasicTabComponent } from '../basic-tab/basic-tab.component';
         slideInAnimation
     ]
 })
-// export class NavTabsComponent implements OnInit, AfterContentInit, AfterViewInit, AfterViewChecked, OnDestroy {
-export class NavTabsComponent extends BasicTabComponent {
+
+export class NavTabsComponent extends TabsGroupComponent implements OnInit, AfterContentInit {
 
     @ContentChildren(TabLinkDirective) tabs!: QueryList<TabLinkDirective>;
+    tabGroup: TabLinkDirective[] = [];
 
-    // @Input() activeTabStyle: string = 'border'; // also 'fill' can be used
-    // @Input() animation: boolean = false;
-    //
-    // @ViewChild('parent', {static: true}) containerElement;
-    // @ViewChild('tabs', {static: true}) tabsWrapperElement;
-    // @ViewChild('arrowLeft', {static: true}) arrowLeftElement;
-    // @ViewChild('arrowRight', {static: true}) arrowRightElement;
-
-    // subscription: Subscription = new Subscription();
-    //
-    // tabsViewElements: HTMLElement[];
-    // curTab: HTMLElement;
-    // tabGroup: TabLinkDirective[] = [];
-    // containerWidth: number;
-    // tabsWrapperWidth: number;
-    // allTabsWidth: number = 0;
-    // arrowWidth: number = 40;
-    // isArrows: boolean;
-    // tabsScrollRect: ClientRect;
-    // curTabClientRect: ClientRect;
-    // containerRect: ClientRect;
-    //
-    // containerPosition$ = new BehaviorSubject(null);
-    //
-    // slideMeasure: { width: any, left: any } = {
-    //     width: 0,
-    //     left: 0
-    // };
+    @ViewChild('parent', {static: true}) containerElement;
+    @ViewChild('tabs', {static: true}) tabsWrapperElement;
+    @ViewChild('arrowLeft', {static: true}) arrowLeftElement;
+    @ViewChild('arrowRight', {static: true}) arrowRightElement;
 
     constructor(
         public themeService: ThemeService,
@@ -93,144 +66,58 @@ export class NavTabsComponent extends BasicTabComponent {
         });
     }
 
-    scrollRight(step = null): void {
-        let x: number = parseInt(this.tabsWrapperElement.style.left, 10) || 0;
-        const defaultStep = this.containerRect.width / 100 * 30;
-        let scrollStep = defaultStep;
-        this.tabsScrollRect = this.tabsWrapperElement.getBoundingClientRect();
-
-        if (this.tabsScrollRect.right - scrollStep < this.containerRect.right + this.arrowWidth) {
-            scrollStep = (this.tabsScrollRect.right - this.containerRect.right) + this.arrowWidth;
-        } else {
-            scrollStep = step ? step : defaultStep;
-        }
-
-        x -= scrollStep;
-
-        this.moveContainer(x);
-    }
-
-    scrollLeft(step = null): void {
-        let x: number = parseInt(this.tabsWrapperElement.style.left, 10) || 0;
-        const defaultStep = this.containerRect.width / 100 * 30;
-        let scrollStep = defaultStep;
-        this.tabsScrollRect = this.tabsWrapperElement.getBoundingClientRect();
-
-        if (this.tabsScrollRect.left + scrollStep > this.containerRect.left + this.arrowWidth) {
-            x = 0;
-        } else {
-            scrollStep = step ? step : defaultStep;
-            x += scrollStep;
-        }
-
-        this.moveContainer(x);
-    }
-
-    moveContainer(x): void {
-        let newX: number = x;
-        if (Math.abs(x) + this.containerWidth - (this.arrowLeftElement.offsetWidth * 2) > this.allTabsWidth) {
-            newX = this.containerWidth - (this.arrowLeftElement.offsetWidth * 2) - this.allTabsWidth;
-        }
-        this.tabsWrapperElement.style.left = newX + 'px';
-        this.containerPosition$.next(true);
-    }
-
-    setUnderlineMeasure(): void {
-        this.curTabClientRect = this.curTab.getBoundingClientRect();
-        this.slideMeasure.width = `${this.curTabClientRect.width}px`;
-        this.slideMeasure.left = `${this.curTab.offsetLeft}px`;
-    }
-
-    setSizes(): void {
-        this.containerRect = this.containerElement.getBoundingClientRect();
-        this.containerWidth = this.containerRect.width;
-        this.tabsWrapperWidth = this.containerWidth - (this.isArrows ? (this.arrowWidth * 2) : 0);
-    }
-
-    setTabSizes(): void {
-        this.allTabsWidth = 0;
-        this.tabsViewElements.forEach(tab => {
-            const tabWidth = tab.offsetWidth;
-            this.allTabsWidth += tabWidth;
-            if (tabWidth > this.tabsWrapperWidth) {
-                tab.classList.add('sdk-tab-container__tab--oversize');
-            }
-        });
-
-        this.tabsScrollRect = this.tabsWrapperElement.getBoundingClientRect();
-    }
-
-    changeRects(): void {
-        this.tabsScrollRect = this.tabsWrapperElement.getBoundingClientRect();
-        this.curTabClientRect = this.curTab.getBoundingClientRect();
-    }
-
     getRouteAnimation(outlet: RouterOutlet): number {
         return outlet.activatedRouteData.index === undefined ? -1 : outlet.activatedRouteData.index;
     }
 
-    // ngOnInit(): void {
-    //     this.containerElement = this.containerElement.nativeElement || this.containerElement;
-    //     this.tabsWrapperElement = this.tabsWrapperElement.nativeElement || this.tabsWrapperElement;
-    //     this.arrowLeftElement = this.arrowLeftElement.nativeElement || this.arrowLeftElement;
-    //     this.arrowRightElement = this.arrowRightElement.nativeElement || this.arrowRightElement;
-    //     this.setSizes();
-    //
-    //     this.subscription = this.containerPosition$.pipe(delay(400)).subscribe(() => this.changeRects());
-    //
-    //     const subResize = fromEvent(window, 'resize')
-    //         .subscribe(() => {
-    //             this.setSizes();
-    //             this.isArrows = this.allTabsWidth > this.containerWidth;
-    //             if (!this.isArrows) {
-    //                 const x = Math.abs(parseFloat(this.tabsWrapperElement.style.left)) || 0;
-    //                 this.tabsWrapperElement.style.left = '0px';
-    //                 if (x !== 0) {
-    //                     this.slideMeasure.left = `${parseFloat(this.slideMeasure.left) + x}px`;
-    //                 }
-    //             }
-    //         });
-    //     const subRouter = this.router.events
-    //         .pipe(
-    //             filter((event: NavigationEnd) => event instanceof NavigationEnd)
-    //         )
-    //         .subscribe(() => {
-    //             this.selectTab();
-    //         });
-    //     const subRightArrow = fromEvent(this.arrowRightElement, 'click')
-    //         .pipe(throttleTime(500))
-    //         .subscribe(() => {
-    //             this.scrollRight();
-    //         });
-    //     const subLeftArrow = fromEvent(this.arrowLeftElement, 'click')
-    //         .pipe(throttleTime(500))
-    //         .subscribe(() => {
-    //             this.scrollLeft();
-    //         });
-    //
-    //     this.subscription.add(subResize);
-    //     this.subscription.add(subRouter);
-    //     this.subscription.add(subRightArrow);
-    //     this.subscription.add(subLeftArrow);
-    // }
+    ngOnInit(): void {
+        this.containerElement = this.containerElement.nativeElement || this.containerElement;
+        this.tabsWrapperElement = this.tabsWrapperElement.nativeElement || this.tabsWrapperElement;
+        this.arrowLeftElement = this.arrowLeftElement.nativeElement || this.arrowLeftElement;
+        this.arrowRightElement = this.arrowRightElement.nativeElement || this.arrowRightElement;
+        this.setSizes();
 
-    // ngAfterContentInit(): void {
-    //     this.tabGroup = [];
-    //     this.tabs.forEach(tabInstance => this.tabGroup.push(tabInstance));
-    //     this.selectTab();
-    // }
+        this.subscription = this.containerPosition$.pipe(delay(400)).subscribe(() => this.changeRects());
 
-    // ngAfterViewInit(): void {
-    //     this.tabsViewElements = Array.from(this.tabsWrapperElement.children);
-    //     this.setTabSizes();
-    //     this.isArrows = this.allTabsWidth > this.containerWidth;
-    // }
-    //
-    // ngAfterViewChecked(): void {
-    //     this.cdRef.detectChanges();
-    // }
-    //
-    // ngOnDestroy(): void {
-    //     this.subscription.unsubscribe();
-    // }
+        const subResize = fromEvent(window, 'resize')
+            .subscribe(() => {
+                this.setSizes();
+                this.isArrows = this.allTabsWidth > this.containerWidth;
+                if (!this.isArrows) {
+                    const x = Math.abs(parseFloat(this.tabsWrapperElement.style.left)) || 0;
+                    this.tabsWrapperElement.style.left = '0px';
+                    if (x !== 0) {
+                        this.slideMeasure.left = `${parseFloat(this.slideMeasure.left) + x}px`;
+                    }
+                }
+            });
+        const subRouter = this.router.events
+            .pipe(
+                filter((event: NavigationEnd) => event instanceof NavigationEnd)
+            )
+            .subscribe(() => {
+                this.selectTab();
+            });
+        const subRightArrow = fromEvent(this.arrowRightElement, 'click')
+            .pipe(throttleTime(500))
+            .subscribe(() => {
+                this.scrollRight();
+            });
+        const subLeftArrow = fromEvent(this.arrowLeftElement, 'click')
+            .pipe(throttleTime(500))
+            .subscribe(() => {
+                this.scrollLeft();
+            });
+
+        this.subscription.add(subResize);
+        this.subscription.add(subRouter);
+        this.subscription.add(subRightArrow);
+        this.subscription.add(subLeftArrow);
+    }
+
+    ngAfterContentInit(): void {
+        this.tabGroup = [];
+        this.tabs.forEach(tabInstance => this.tabGroup.push(tabInstance));
+        this.selectTab();
+    }
 }
