+++ libs/ngx-slice-kit/src/lib/navigation/tabs-group/tabs-group.module.ts
@@ -1,15 +1,15 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { BasicTabComponent } from './basic-tab.component';
 import { IconModule } from '../../buttons/icon/icon.module';
+import { TabsGroupComponent } from './tabs-group.component';
 
 
 @NgModule({
     declarations: [
-        BasicTabComponent
+        TabsGroupComponent,
     ],
     exports: [
-        BasicTabComponent
+        TabsGroupComponent
     ],
     imports: [
         CommonModule,
@@ -20,5 +20,5 @@ import { IconModule } from '../../buttons/icon/icon.module';
         NO_ERRORS_SCHEMA,
     ]
 })
-export class BasicTabModule {
+export class TabsGroupModule {
 }
